requirements.txt
iniconfig==2.1.0
python-dotenv==1.1.1
patchright==1.52.0
playwright==1.52.0

